+++ src/TextHeatmaps.jl
@@ -6,13 +6,14 @@ using Colors: Colorant, RGB, hex
 using ColorSchemes: ColorScheme, colorschemes, get
 using XAIBase: Attribution, AbstractXAIMethod, analyze
 using XAIBase: AbstractTransform, Pipeline
+import XAIBase: compose
 using XAIBase: Batch, eachsample, mapsamples
-using XAIBase: AbstractPooling, UnsignedPooling, SignedPooling, pool
+using XAIBase: AbstractPooling, pool
 using XAIBase: SumPooling, MaxPooling, SignedNoPooling, UnsignedNoPooling
 using XAIBase: SumAbsPooling, AbsSumPooling, MaxAbsPooling, NormPooling, SquaredNormPooling
 using XAIBase: AbstractNormalization, ExtremaNormalization, CenteredNormalization
 using XAIBase: BatchedNormalization
-using XAIBase: normalize, default_normalization
+using XAIBase: normalize, default_normalization, issigned
 
 include("transforms.jl")
 export Colormap